+++ scripts/constants/materials.gd
@@ -1,0 +1,7 @@
+class_name Materials
+
+const RANGE_BUILD: Material = preload("res://assets/materials/indicators/range_build.tres")
+const RANGE_CITY_HIGHLIGHT: Material = preload("res://assets/materials/indicators/range_city_highlight.tres")
+
+const RANGE_DEFENSE: Material = preload("res://assets/materials/indicators/range_defense.tres")
+const RANGE_DEFENSE_UNSUPPLIED: Material = preload("res://assets/materials/indicators/range_defense_unsupplied.tres")